Move RecipeMapBackend to a separate file
