+++ src/main/java/gregtech/api/recipe/RecipeMapBackend.java
@@ -20,8 +20,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.Spliterators;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
@@ -57,7 +55,6 @@ import gregtech.api.recipe.lookup.GTOreDictLookupIngredient;
 import gregtech.api.recipe.lookup.GTRecipeLookup;
 import gregtech.api.recipe.lookup.GTRecipeLookupBuilder;
 import gregtech.api.recipe.lookup.GTRecipeLookupIngredient;
-import gregtech.api.util.GTLog;
 import gregtech.api.util.GTOreDictUnificator;
 import gregtech.api.util.GTRecipe;
 import gregtech.api.util.GTRecipeBuilder;
@@ -453,7 +450,8 @@ public class RecipeMapBackend {
     Stream<GTRecipe> matchRecipeStream(@Nullable ItemStack @NotNull [] rawItems,
         @Nullable FluidStack @NotNull [] fluids, @Nullable ItemStack specialSlot, @Nullable GTRecipe cachedRecipe,
         boolean notUnificated, boolean dontCheckStackSizes, boolean forCollisionCheck) {
-        RecipeLookupProfile profile = RecipeLookupProfile.start(recipeMap, forCollisionCheck);
+        RecipeMapBackendProfiler.RecipeLookupProfile profile = RecipeMapBackendProfiler.RecipeLookupProfile
+            .start(recipeMap, forCollisionCheck);
         long setupStart = profile == null ? 0L : System.nanoTime();
         if (doesOverwriteFindRecipe()) {
             if (profile != null) {
@@ -1327,7 +1325,7 @@ public class RecipeMapBackend {
 
     private Stream<GTRecipe> collisionCandidateStream(@Nullable ItemStack @NotNull [] items,
         @Nullable FluidStack @NotNull [] fluids, @Nullable ItemStack specialSlot, boolean dontCheckStackSizes,
-        @Nullable RecipeLookupProfile profile) {
+        @Nullable RecipeMapBackendProfiler.RecipeLookupProfile profile) {
         Stream<GTRecipe> candidates = lookupCandidateStream(items, fluids, profile);
         if (profile != null) {
             candidates = candidates.peek(recipe -> profile.recordTrieCandidate());
@@ -1340,7 +1338,7 @@ public class RecipeMapBackend {
     }
 
     private Stream<GTRecipe> lookupCandidateStream(@Nullable ItemStack @NotNull [] items,
-        @Nullable FluidStack @NotNull [] fluids, @Nullable RecipeLookupProfile profile) {
+        @Nullable FluidStack @NotNull [] fluids, @Nullable RecipeMapBackendProfiler.RecipeLookupProfile profile) {
         if (profile == null) {
             return lookupCandidateStream(items, fluids);
         }
@@ -1450,7 +1448,7 @@ public class RecipeMapBackend {
 
     private boolean profiledFilterFindRecipe(@NotNull GTRecipe recipe, @Nullable ItemStack @NotNull [] items,
         @Nullable FluidStack @NotNull [] fluids, @Nullable ItemStack specialSlot, boolean dontCheckStackSizes,
-        @Nullable RecipeLookupProfile profile) {
+        @Nullable RecipeMapBackendProfiler.RecipeLookupProfile profile) {
         if (profile == null) {
             return filterFindRecipe(recipe, items, fluids, specialSlot, dontCheckStackSizes);
         }
@@ -1463,7 +1461,7 @@ public class RecipeMapBackend {
     @Nullable
     private GTRecipe profiledModifyFoundRecipe(@NotNull GTRecipe recipe, @Nullable ItemStack @NotNull [] items,
         @Nullable FluidStack @NotNull [] fluids, @Nullable ItemStack specialSlot,
-        @Nullable RecipeLookupProfile profile) {
+        @Nullable RecipeMapBackendProfiler.RecipeLookupProfile profile) {
         if (profile == null) {
             return modifyFoundRecipe(recipe, items, fluids, specialSlot);
         }
@@ -1474,277 +1472,6 @@ public class RecipeMapBackend {
     }
 
     // endregion
-
-    private static final class RecipeLookupProfile {
-
-        private static final boolean ENABLED = Boolean.getBoolean("gt.recipe.lookup.profile");
-        private static final long REPORT_INTERVAL_CALLS = Long
-            .getLong("gt.recipe.lookup.profile.interval_calls", 5_000L);
-        private static final long REPORT_INTERVAL_NANOS = Math
-            .max(1L, Long.getLong("gt.recipe.lookup.profile.interval_seconds", 15L)) * 1_000_000_000L;
-        private static final ConcurrentMap<String, RecipeLookupProfileStats> STATS = new ConcurrentHashMap<>();
-
-        private final RecipeLookupProfileStats stats;
-
-        private RecipeLookupProfile(RecipeLookupProfileStats stats) {
-            this.stats = stats;
-        }
-
-        @Nullable
-        private static RecipeLookupProfile start(@Nullable RecipeMap<?> recipeMap, boolean collisionCheck) {
-            if (!ENABLED) {
-                return null;
-            }
-            String mapName = recipeMap == null ? "<unbound>" : recipeMap.unlocalizedName;
-            RecipeLookupProfileStats stats = STATS.computeIfAbsent(mapName, RecipeLookupProfileStats::new);
-            stats.recordCall(collisionCheck);
-            return new RecipeLookupProfile(stats);
-        }
-
-        private void recordOverwrite() {
-            stats.recordOverwrite();
-        }
-
-        private void recordEmptyMapReject() {
-            stats.recordEmptyMapReject();
-        }
-
-        private void recordMinInputReject() {
-            stats.recordMinInputReject();
-        }
-
-        private void addSetupNanos(long nanos) {
-            stats.addSetupNanos(nanos);
-        }
-
-        private void addUnificationNanos(long nanos) {
-            stats.addUnificationNanos(nanos);
-        }
-
-        private void addEnsureLookupNanos(long nanos) {
-            stats.addEnsureLookupNanos(nanos);
-        }
-
-        private void recordCachedRecipeCandidate() {
-            stats.recordCachedRecipeCandidate();
-        }
-
-        private void recordCacheMapProbe() {
-            stats.recordCacheMapProbe();
-        }
-
-        private void recordCacheMapCandidate() {
-            stats.recordCacheMapCandidate();
-        }
-
-        private void recordTrieLookupSetup() {
-            stats.recordTrieLookupSetup();
-        }
-
-        private void addTrieLookupSetupNanos(long nanos) {
-            stats.addTrieLookupSetupNanos(nanos);
-        }
-
-        private void recordTrieCandidate() {
-            stats.recordTrieCandidate();
-        }
-
-        private void recordMatchedCandidate() {
-            stats.recordMatchedCandidate();
-        }
-
-        private void recordFallbackProbe() {
-            stats.recordFallbackProbe();
-        }
-
-        private void recordFallbackHit() {
-            stats.recordFallbackHit();
-        }
-
-        private void addFilterNanos(long nanos, boolean matched) {
-            stats.addFilterNanos(nanos, matched);
-        }
-
-        private void addModifyNanos(long nanos, boolean returnedRecipe) {
-            stats.addModifyNanos(nanos, returnedRecipe);
-        }
-    }
-
-    private static final class RecipeLookupProfileStats {
-
-        private final String mapName;
-        private long lastReportNanos = System.nanoTime();
-        private long lastReportCalls;
-        private long calls;
-        private long collisionCalls;
-        private long overwriteCalls;
-        private long emptyMapRejects;
-        private long minInputRejects;
-        private long setupNanos;
-        private long unificationNanos;
-        private long ensureLookupNanos;
-        private long cachedRecipeCandidates;
-        private long cacheMapProbes;
-        private long cacheMapCandidates;
-        private long trieLookupSetups;
-        private long trieLookupSetupNanos;
-        private long trieCandidates;
-        private long matchedCandidates;
-        private long fallbackProbes;
-        private long fallbackHits;
-        private long filterCalls;
-        private long filterMatches;
-        private long filterNanos;
-        private long modifyCalls;
-        private long modifyHits;
-        private long modifyNanos;
-
-        private RecipeLookupProfileStats(String mapName) {
-            this.mapName = mapName;
-        }
-
-        private synchronized void recordCall(boolean collisionCheck) {
-            calls++;
-            if (collisionCheck) {
-                collisionCalls++;
-            }
-            maybeReport();
-        }
-
-        private synchronized void recordOverwrite() {
-            overwriteCalls++;
-        }
-
-        private synchronized void recordEmptyMapReject() {
-            emptyMapRejects++;
-        }
-
-        private synchronized void recordMinInputReject() {
-            minInputRejects++;
-        }
-
-        private synchronized void addSetupNanos(long nanos) {
-            setupNanos += nanos;
-        }
-
-        private synchronized void addUnificationNanos(long nanos) {
-            unificationNanos += nanos;
-        }
-
-        private synchronized void addEnsureLookupNanos(long nanos) {
-            ensureLookupNanos += nanos;
-        }
-
-        private synchronized void recordCachedRecipeCandidate() {
-            cachedRecipeCandidates++;
-        }
-
-        private synchronized void recordCacheMapProbe() {
-            cacheMapProbes++;
-        }
-
-        private synchronized void recordCacheMapCandidate() {
-            cacheMapCandidates++;
-        }
-
-        private synchronized void recordTrieLookupSetup() {
-            trieLookupSetups++;
-        }
-
-        private synchronized void addTrieLookupSetupNanos(long nanos) {
-            trieLookupSetupNanos += nanos;
-        }
-
-        private synchronized void recordTrieCandidate() {
-            trieCandidates++;
-        }
-
-        private synchronized void recordMatchedCandidate() {
-            matchedCandidates++;
-        }
-
-        private synchronized void recordFallbackProbe() {
-            fallbackProbes++;
-        }
-
-        private synchronized void recordFallbackHit() {
-            fallbackHits++;
-        }
-
-        private synchronized void addFilterNanos(long nanos, boolean matched) {
-            filterCalls++;
-            if (matched) {
-                filterMatches++;
-            }
-            filterNanos += nanos;
-        }
-
-        private synchronized void addModifyNanos(long nanos, boolean returnedRecipe) {
-            modifyCalls++;
-            if (returnedRecipe) {
-                modifyHits++;
-            }
-            modifyNanos += nanos;
-        }
-
-        private void maybeReport() {
-            long now = System.nanoTime();
-            boolean enoughCalls = RecipeLookupProfile.REPORT_INTERVAL_CALLS > 0
-                && calls - lastReportCalls >= RecipeLookupProfile.REPORT_INTERVAL_CALLS;
-            boolean enoughTime = now - lastReportNanos >= RecipeLookupProfile.REPORT_INTERVAL_NANOS;
-            if (!enoughCalls && !enoughTime) {
-                return;
-            }
-
-            GTLog.out.println(
-                String.format(
-                    Locale.ROOT,
-                    "[GTRecipeLookupProfile] map=%s calls=%d collision=%d overwrite=%d empty=%d minReject=%d "
-                        + "setupAvgUs=%.3f ensureAvgUs=%.3f unifyMs=%.3f cachedCandidates=%d "
-                        + "cacheMapProbes=%d cacheMapCandidates=%d trieLookupSetups=%d trieLookupSetupMs=%.3f "
-                        + "trieCandidates=%d trieCandidatesPerLookupSetup=%.3f matched=%d fallbackProbes=%d fallbackHits=%d "
-                        + "filterCalls=%d filterMatches=%d filterMs=%.3f modifyCalls=%d modifyHits=%d modifyMs=%.3f",
-                    mapName,
-                    calls,
-                    collisionCalls,
-                    overwriteCalls,
-                    emptyMapRejects,
-                    minInputRejects,
-                    averageMicros(setupNanos, calls),
-                    averageMicros(ensureLookupNanos, calls),
-                    nanosToMillis(unificationNanos),
-                    cachedRecipeCandidates,
-                    cacheMapProbes,
-                    cacheMapCandidates,
-                    trieLookupSetups,
-                    nanosToMillis(trieLookupSetupNanos),
-                    trieCandidates,
-                    average(trieCandidates, trieLookupSetups),
-                    matchedCandidates,
-                    fallbackProbes,
-                    fallbackHits,
-                    filterCalls,
-                    filterMatches,
-                    nanosToMillis(filterNanos),
-                    modifyCalls,
-                    modifyHits,
-                    nanosToMillis(modifyNanos)));
-            lastReportCalls = calls;
-            lastReportNanos = now;
-        }
-
-        private static double average(long value, long count) {
-            return count == 0 ? 0.0D : (double) value / count;
-        }
-
-        private static double averageMicros(long nanos, long count) {
-            return count == 0 ? 0.0D : (double) nanos / count / 1_000.0D;
-        }
-
-        private static double nanosToMillis(long nanos) {
-            return (double) nanos / 1_000_000.0D;
-        }
-    }
 
     @FunctionalInterface
     public interface BackendCreator<B extends RecipeMapBackend> {
